requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# API Framework
FastAPI==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy>=2.0.0

# AI/ML
langchain>=0.3.0
langchain-groq>=0.1.0
langchain-huggingface>=0.0.1
langgraph>=0.2.0

# RAG & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# Testing (dev)
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0

# Code Quality (dev)
black==24.10.0
isort==5.13.2
ruff==0.7.0

# Frontend
streamlit>=1.28.0
requests>=2.31.0
plotly>=5.17.0

# Utilities
httpx==0.24.0
